Refactor to move createGeometryFromParams into Spatial Service
